Logix_WorkingModel/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-generativeai
pydantic
pathway